Moved trace_version helper.
