+++ zcash_local_net/src/logs.rs
@@ -56,20 +56,3 @@ impl<T: LogsToDir> LogsToStdoutAndStderr for T {
         print_log(stdout_log_path);
     }
 }
-
-/// Helper to trace the executable version.
-pub fn trace_version(executable_name: &str, version_command: &str) {
-    let mut command = crate::utils::executable_finder::pick_command(executable_name);
-
-    let args = vec![version_command];
-
-    command.args(args);
-
-    let version = command
-        .output()
-        .expect(crate::utils::executable_finder::EXPECT_SPAWN);
-    tracing::info!(
-        "$ {executable_name} {version_command}
- {version:?}"
-    );
-}
